Pass key into `rewriteLeaf`.
To begin to track the key of a leaf page in the positions entry, we need
to pass it into `writePositions`. `writePositions` is called in two
places, `rewriteLeaf` and `exorcise`. `writePositions` will eventually
accept an array that contains the key, or that is empty if there is no
key. This commit prepares the implementation of `writePositions` by
providing `writeLeaf` with the key.

See #135.