Cleanup dead code for JSON record reads.
I'm using the length of the record stored in the positions entry to read
a record not loaded when the leaf page log is played, but I've left the
code that scans for new line characters and the code that complains when
a new line character cannot be found. This is dead code and it has been
removed.

See #154.