Remove record count added to each log entry.
I need to make way for the addresses in insert record entries, so it's
time to tidy them by removing the record count that I've been storing in
each entry. The record count is currently unused and ultimately destined
for the leaf page footer record. I want to remove it before I add the
address to branch page insert entries.

Closes #163.