Explicitly set key when ghost busting.
Clearing the keys meant that the new key would be designated after all
the locks had been released during the next descent. The first pass at
our new logic looked to promote a key in the page passed to `exorcise`,
but in the case of a merge of an empty left leaf page to a right leaf
page that is not empty, that promotion is going to fail. It is going to
fail because there is nothing to promote in that page because that page
is empty.

Now `exorcised` needs a pivot branch page, a leaf page who ghost needs
busting, and a page from which to obtain a key, accounting for a ghost
in that key.

Closes #186.