`deleteGhost` not returns key.
Fixed a bug in `deleteGhost` and `splitLeaf`. Instead of returning the
new key, `deleteGhost` was returning the position of the key record.
This value is only used in `splitLeaf` where it updates the key used to
search for the leaf page to split.

There is only one test that uses that particular branch of code,
`t/ghosts/split.t.js`. I'd rather the test failed, but it survived.

I'd rather it failed when it got a key value of 104 instead of 'e', but
we're already at the penultimate branch, so we never end up using the
bad, bad key to descent. We're already where we need to be. If the tree
used in `t/ghosts/split.t.js` had three levels, it would fail.

Closes #193.