Ignore subsequent right peek from leaf page merge.
Ignore the results of peeking at the subsequent right leaf page, the
right sibling leaf page of the right leaf page of a leaf page merge.

When merging leaf pages, we were peeking at the right sibling leaf page
of the right leaf page in the merge for a key to replace the pivot key.
If we do need to replace the pivot key, we can obtain a new key from the
ancestor branch page instead of performing this peek. The same logic
works for both leaf page merges and branch page merges.

This has been discussed in the commit messages just prior to this one.

See #187.