Remove last of leaf page peek pivot replacement.
Remove the last of the logic for a replacement of a pivot key using the
key of the right sibling leaf page of the right leaf page of a leaf page
merege. We now extract our keys for merge pivot key replacement from the
branch pages, so the logic is the same for leaf page merges as it is for
branch page merges.

Closes #187.