Designate on load.
As an interim step toward storing keys in leaves, determine the key
values on load. Looks as though I'm not going to designate the left most
child of a branch page. It would mean that I'd have to rewrite all the
leaves along the left edge of a sub-tree.

The next action is to pass keys in to the `chooseBranchToMerge` function
and to track the left and right pivots as I descend to the branch to
merge.

Closes #196.