Pass key into `chooseBranchesToMerge`.
Obtaining a key to pass into `chooseBranchesToMerge`. The key is a key
that is or was in the `ancestor` branch page, one that will send the
descent through the branch page, however the address is still required
in order to stop the descent.

The key is obtained from the branch pages of the merge logic, without
descending to a leaf page to obtain a key.

See #188.