Eliminate `designate`.
I've eliminated `designate`, the function that would load keys for
branches by descending to the leaf page to obtain the key. It continues
to exist since I still use it to load the keys when the branch pages
loads, but that is only temporary until branch page key journaling is
implemented.

Closes #189.