Assert address of right branch page in merge.
Assert that the address of the right branch page in a merge is the
address sought when `mergeBranches` is called. When `mergeBranches` is
called it is called with the address of the right branch page of the
merge and the key for the left most child of the right branch page,
determined by descending to the left most leaf page beneath the right
branch page.

See #201.