Call `_designate` if necessary in `find`.
Trying to phase out `_designate` but I need to call it for now, but only
when I've neglected to update the cache during a split or merge, so I've
placed the call to designate in a separate branch so I can see when it
is called with debugging statements, even though `_designate` itself
will check the cache.

See #201.
See #189.