Diagnose leaf page merge exorcism bug.
Checking in a day of re-orientation after having been away for so long.

Found that I'm not correctly determining the parent branch page of the
left leaf page of a page merge so that if the page merge requires an
exorcism, I'm not providing the correct parent of the left leaf page for
the exorcism.

When we were just clearing keys, clearing the wrong key didn't matter,
it would get read in at the next descent. The locking was still correct
even though we were clearing the wrong key.

How was it that this was not caught by existing tests?

Because we descend to the left by going always to the right, we weren't
checking keys and thus we were not reading in keys. In the root filling
test of the merge tests, the test that I'm using to diagnose the
problem, there is a ghost busting leaf page merge, but the previous
operation in the balance plan cleared the key for the parent of the left
branch page. Because the descent to the left page does not use the key,
the change of key in the left page causes no confusion. The key of it's
parent is not in the cache. It is cached on a subsequent read, after the
merge and change of key are complete.

See #202.