Track singles on descent to left page to merge.
With this commit, I'm tracking singles on the descent to the left page
to merge, simply tracking, not asserting that I'm using the ghosted page
that I've derived from the tracking correctly.

Noticing that `Descent.indexes` is only used by page merge. If that is
the case, then maybe the custom `unlocker` can provide both the page and
the index and `mergePages` can track this instead of the descent.

See #203.
See #202.