Set key in middle of merged page.
Use the key given by `chooseBranchToMerge` to `mergeBranches` as the key
for the left most child of the right branch page. The left most child of
any branch page does not have a key in the branch page key cache. It
represents the values that are less than the key of the second child
from the left. The key is used as the key of a child in a branch page
above the right branch pages to merge. It is moved from its position as
a key above the right branch page to a key within the newly merged
branch page. The pivot is replaced by the generalized `mergePages`
function.

The previous commit corrected the error were the wrong branch page was
given as the keyed ancestor of the left branch page. With this commit we
see that it works correctly, the correct ancestor is rekeyed when a
merge exorcises a ghost record. This wrong keyed ancestor error was
detected by `t/merge/root-fill.t.js`.

Closes #202.
See #201.