Expose `address` and `right` address in Cursor.
It's a race condition when we go backwards by going to the left of the
current page. In the time it takes for us to release our cursor and
descend the tree again the tree could balance, eliminating the key we're
using to go left.

Checking that the right sibling matches the address of the page we just
left will tell us if we've won the race, otherwise we retry.

Closes #247.