Move right page address to footer.
Move the right page address into the footer and remove it from the
positons array. Leave the ghosts in the position array so that we can
replay the log from the positions array using the ghosts count at the
time the position array was written. Indeed, the ghosts count is a
property of the positions array in a leaf page, it is a partition of the
array and it should be stored with the array.

Closes #222.