Use page read/write lock to guard page load.
Simplify page loading by using the same read/write lock that guards the
page to guard the loading of the page. Instead of a `loaded` property
for the page, we now add a level of indirection to the value of the
Magazine cache entry. The value is not the page, but an object that
contains the page in a page property. When it is missing, the page has
not been loaded. If we fail to we clear the property and propagate the
error to any descents waiting in the Sequester read/write lock the page
object that did not load.

Closes #266.