Test `readRecord` error handling, fix janitor.
The `janitor` function given to `validator` was lost. It was not part of
the parameters to `validator`, but incorrectly added to the parameters
of the function that `validator` creates.

The `janitor` is not run if the error is an exception thrown from user
land.

Test the error branch in read record that skips adding the record to the
page's record cache, seeing as how the record does not exist.

Closes #259.
See #263.