Pass `Descent` into `Cursor`.
Take a step toward the structured hand-off of the `Locker` by passing
the `Descent` into the `Cursor` and having the cursor zero out an array
that contains the `Descent`. We'll use this as a reference to determine
if the `Cursor` owns the `Descent` or if the `Descent` is still owned by
the `Strata` object.

See #273.