Test merge where left parent has single child.
Test a merge of leaf pages where the left sibling contains a ghost and
it is the only child of its parent meaning that the ghost manipulation
needs to take place in its grand-parent.

Also, fixed `stratify.bin.js` so it will delete an entry only if the
entry exsits. The `if` statement was in the wrong place.

Closes #209.