Upgrade Proof to 0.0.50.
Proof 0.0.50 has a new scaffold. This called for a change to the test
harness and all off the tests. Proof no longer provides cleanup, but the
test harness, the proof wrapper in Strata, it will now perform the
cleanup before and after.

The common utility functions are now exported globally. They may move
into a utility object at some point in the future.

Closes #411.