Update referring leaf page when ghost is deleted.
This is the beginnging of leaf rotation. I've created a script object
that will be used to journal structural changes to the tree. Alls
structural changes will be implemented using the script object. The
script object can be tested separately and the flakey file system
conditions tested without having to descend a tree.

The count of ghosts is now written to the header, when it is zero, we
know to shift the ghost if there is one when replaying a log.

See #453.
See #444.
See #413.