Sketch of synchronous descent.
Or at least more synchronous than the previous descent. Instead of
maintaining callbacks that block others from descending we try to get to
the leaf without having to load a page but give up and start over if do
have to load a page. We are using a Cadence loop to descend but with
fewer asynchronous calls to locking mechanisms.