Sketch of a commit utility.
All rewriting of branches for merge can be expressed with splice an
unlink. We'll need a special split operation for splits. We can make our
updates in memory synchronously, write the writes that were pending,
then quickly write out the rebalance as a plan and allow writing to
continue. We hold the pages in memory, so that they do not need to be
loaded while we are committing to our plan. Regular writes to leaves can
continue while the branches are being rewritten.