Read pages for branch commit, update heft.
We want the pages as they exist on the file system during commit. In
memory they are already going to be altered for the splits, except when
we are re-running a failed commit at startup. We load the page from the
cache after we write it because of startup recovery, we're sure to load
the page as it was just written.