Where I left off notes, alpahbet generators.
Created an alphabet generator to create words to use to create larger
trees for splitting. Integers are very easy to generate, of course, but
I find it easer to read an alphabet for some reason, maybe to as not
confuse with addresses.

Lots of notes on where I left off, loading the project back into my
head, hopefully making it easier to reload in the future.

Closes #544.