Search for preceding value, missed fork.
Using fork to find the previous page using the first key of the current
page, but fork returns null if the key is not actually used as a page
key, if it doesn't exist in the branch pages. This is used by select
merger.  Before you think it is not necessary since the left-most leaf
page is known, and therefore the page left of it should not be sought,
we need this logic for inspecting the the branch page left of a branch
page.

When using fork to iterate we have a race condition because the page may
be merged and the key used to descend to the left page removed from the
tree. I've updated descend to take an `approximate` flag that will
return the value before the given key. If it is not used as a page key,
then we'll simply move back one when we descend the tree and land on the
key within a page, or return the resolved index if we descend the tree
and the key is not there at all.

Closes #581.