Allow binary keys.
Serialize the key as unknown binary data the same way you serialize the
body. Now instead of having a key and body, the serialized entry has an
array of zero or more parts. No longer makes sense to have the player do
the JSON conversion. Conversion should be a pair of serialize and
deserialize functions, no-ops for `Buffer`. I sketched out how you'd
specify that in `Journalist`.

Closes #586.