Implement a null cursor.
A lot of looping assumes that the loop starts with an open cursor that
could be resused, if not a new search is performed. This implements a
bogus open cursor that will always return null from `Cursor.indexOf`
indicating that the value will not be found in the current bogus page
and the caller should open a new cursor.

Closes #595.