Minimum heft of 1, heft sums review.
Ensure that a leaf page has a mimimum heft of `1` so that a cache purge
to zero will remove all the pages if all pages are released. If the
minimum heft is `0` and there are multiple pages at the head of the list
that are `0` they will not be removed, we will have already reached
zero. Think about it.

Added some assertions to check that we are maintaining `heft` correctly.
They should be removed in time.

Closes #603.