Add an instance to the cache key.
Allows the opening and closing of a Strata. Even if we say that the user
is not allowed to reuse a `Strata`, we have already decided that they
can reuse a `Cache`, so they could be dutifully constructing a new
`Strata` for a directory and then pulling pages from a previous cache
instance.

Not that it really should matter, oh, well I suppose it does. Imagine a
user shutting down a `Strata` so another process can use it, update it.
Then opening it up again and using `Cache` entries that are stale? Why?
Why not? How can you say that is an unreasonable expectation?

Not sure why we haven't just resurrected Magazine. It has the ability to
know whether a `Strata` is closing with all its entries released.

Closes #604.