Fix `_id` value, housekeeping key, blocking.
Wasn't correctly incrementing the `_id` in `Journalist.create` so that
if I used a newly created `Strata` to the point where there was a page
split, the new leaf page could get a `0.1` id which already exists and
thereby sew confusion. It really set vacuum in a tizzy, so I'm leaving
some debugging statements I can use to explore that further with Locket
unit tests driving the exploration.

Using the user's key for the housekeeping queue key didn't work because
the user's key can be an arbitrary object. Works fine in the unit tests,
because they key is just a string. Now serializing the key using the
user's key to buffers serialization, then converting the buffers to base
64, then converting the array of base 64 to JSON. This is okay, I guess.

Blocking didn't work because I was deleting the queue entry which allows
a user append to just create a new queue entry and slip past the blocks
into the append implementation. Just delete the queue entry after the
blocks are released.

Renamed the `entries` array for a page to `vacuum` since those records
are the vacuum relevant records and `entries` is used as a local
variable name to track gathered page cache entries for release. Makes it
easier to search the source for the vacuum accounting.

Started typing out some basic Docco in addition to Docco specific to
these changes.