Remove unlocker.
Have the user perform `Cursor.indexOf` at the outset. Now the b-tree is
that much more primitive. The unlocker wasn't working. The `search`
method would awake from an await and see that the page is correct, then
return a cursor in an unlocker, but no telling how long the user will
wait before awaiting the returned from search. By the time they do, the
page may have changed. The only way to ensure that you have the correct
page is to perform tests synchronously.

Closes #612.
Closes #616.