Store key in leaf pages.
Go ahead and store the key in the leaf page instead of trusting that the
first entry has the key. We want to group pages by a subset of a
compound key, so now we have to have a key that is at the start of that
subset, which might not yet exist in the tree.

That is, if we have a string and a number, and the string is the subset,
the key for the page should be the string and the least value of the
number, perhaps zero. If we split on a record with a string 'a' and a
number 3, and we then insert a string 'a' with a number 2, the record
belongs at the zero element of the items array. To date we assume that
we will never insert into the zero position unless we are in the left
most branch.

Closes #656.