Resolve heft races, serializing empty records.
When there are no parts for a 'dependent' record we do not want to give
the parts to our user serializer. We know that the parts are empty, they
do not have to be converted, we'll know to ingore them based on the
header type.

Also, we are using the page key now to find a page for housekeeping, but
the left most page has a null key, so we still have to use the least
value in the page for the left most page.

Now that we calculate heft when we insert a record, a heft race
condition we anticipated has not become a reality. Adjusting the heft
in leaf split meant waiting for any outstanding writes to flush because
heft was calculated on write, but while those writes are flushing heft
can be updated by the user by adding more records. We now adjust heft in
the synchronous period of `_splitLeaf`.

See #661.
See #656.