Fix write leaf data race.
We remove the queue entry for an append in the async `_append` before we
return the write which starts a race. The user may append new records
which would be enqueued in a new queue entry and possibly pulled into
a different turnstile to so that two leaf writes occur simultaneously.

Closes #674.