Propagate block in `_append`.
`_append` will wait on a block. Since we're not immediately deleting the
entry in the queue, we need to immediately delete the block. Then while
after writing the leaf page we need to see if a new block has arrived
and assign it to a new queue entry after deleting the queue entry for
the `_append`.

See #674.