Approximate fork at zero index is null.
When an approximate fork is off of the zero index return `null`. This is
an issue introduced by removing ghosts and recording leaf page keys. It
used to be the case that the zero index would always be the key, so you
couldn't land there if you did an approximate fork, you'd get an actual
fork.

Closes #681.