Fix page read race condition.
Race condition where two strands read a page at the same time and
experience a cache miss at the same time. Goes like this.

The first strand descends synchronously, gets a miss. The miss is
indicated by holding the page in the cache, but the page value for the
entry is null, the page was not in the cache so a new entry was created.
This entry will not be set by the synchronous descent so the synchronous
descent removes the entry. The read process for the first strand begins
by holding the entry which is not going to exist so a new is created
with a null value. It then reads asynchronously.

Okay, now switch to the second strand. It descends the tree and it also
gets a miss, but instead of creating an entry, it receives the entry
created by the read with the null value. Seeing the null value the
second strand's synchronous descent removes the entry, but there are two
references and an assertion that there must only be one. Assert fails.

With this implementation we don't hold onto the entry before reading.
Kind of obvious. And then check that no one else had read the page and
set the cache when we return.

Added a unit test to perform the race.

Closes #732.