Optionally externalize part serialization.
Looking for a limiting value for in-memory stages for Memento. Within
Strata we use `heft`, the serialized byte length of a record. Heft is
determined at insert or remove. It is the return value of insert or
remove. In order to obtain a heft you have to put an item in a tree.

Memento's in-memory stages are arrays. We could use an array count as
our limiting value -- flush when the in-memory stage has a certain
number of elements -- but we'd rather have it based on the size of those
elements so we want to serialize to get a byte length. It will still be
representative, representative of less than half the actual size of an
element in memory since a serialized form will still require a
deserialized form for use in searching and sorting.

Here we externalize a partial serialization returning the array of
buffers that compose an entry body. This gives us a partial heft. The
heft once inserted into the tree will include the byte length of the
header. In the tree it is an approximation of size relative to the sizes
of other elements. The externalized partial serialization can be used as
similarly.

We include the header because the delete values have a zero byte length
for their body. A delete is recorded entirely in the header using the
index of the record in the page as specify which record to delete
instead of a key or any other user specified material.

In Memento a delete is a record added to the page that represents a
deletion in a change log, so externalized without the header deletes
will still have a heft.

Closes #733.