Test housekeeping race.
We have a race where we add a page to housekeeping for split or merge
but the split or merge condition is no longer valid when housekeeping
begins. We test it here by simply reaching split then unreaching split
hopefully before the housekeeping queue gets to work on the entry. Our
hopes are realized with this test, but it may or may not be
deterministic. It may be deterministic, based on the order of entries in
the PromiseJobs queue, or it might be non-deterministic, based on the
vagaries of my workstation. Time will tell. Better coverage than before.

Closes #734.