Remove unnecessary `async` from write-ahead.
Starting to notice that WriteAheadOnly is asynchronous and most of the
writing takes place in background in the Turnstile. Suggests some sort
of refactor where the `'write'` key of the Turnstile or the Turnstile
itself is moved into FileSystem.

At this point I can't see how to move balancing into the storage
stragies without creating a common super-class. Basically, some major
refactor that isn't coming to me in flash, that I can't get excited
about, and seems to suggest some other tangle of classes that is going
to be no better than what we have here which at least I'm familiar with.

The drawback of this implementation is that we await calls to
`WriteAheadOnly` when it is not `async`. After a refactor the draw back
would be having to pass around a `Descender` or abstracting the balance
to with closures, or something along those lines.