Fix comparison against `null`.
When selecting a sibling for merge if the sibling is a leaf we check to
see if the first key matches the last key to see if it is a parition
filled with a partial key value. Except as I type this I realize this is
a bogus test. Anyway, I am not performing the test for branch pages any
longer because a branch can have a null key in the first position.